style: CheckStyle, SpotBugs 적용
