+++ order/src/main/java/project/swithme/order/common/annotation/auth/LoginUser.java
@@ -8,4 +8,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE, ElementType.PARAMETER})
 public @interface LoginUser {
+
 }
