+++ order/src/main/java/project/swithme/order/common/configuration/business/PayConfiguration.java
@@ -1,10 +1,9 @@
 package project.swithme.order.common.configuration.business;
 
+import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import project.swithme.order.core.domain.order.entity.PayGroup;
-
-import java.util.List;
 
 @Configuration
 public class PayConfiguration {
@@ -12,8 +11,8 @@ public class PayConfiguration {
     @Bean
     public PaymentGroup payTypes() {
         List<PayGroup> payGroups = List.of(
-                PayGroup.CASH,
-                PayGroup.CARD
+            PayGroup.CASH,
+            PayGroup.CARD
         );
         return new PaymentGroup(payGroups);
     }
