+++ order/src/main/java/project/swithme/order/common/configuration/caffeine/CaffeineConfiguration.java
@@ -1,13 +1,12 @@
 package project.swithme.order.common.configuration.caffeine;
 
+import static java.util.concurrent.TimeUnit.DAYS;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.caffeine.CaffeineCacheManager;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import static java.util.concurrent.TimeUnit.DAYS;
 
 @EnableCaching
 @Configuration
@@ -16,7 +15,7 @@ public class CaffeineConfiguration {
     @Bean
     public Caffeine<Object, Object> caffeineConfig() {
         return Caffeine.newBuilder()
-                .expireAfterWrite(1, DAYS);
+            .expireAfterWrite(1, DAYS);
     }
 
     @Bean
