+++ order/src/main/java/project/swithme/order/common/configuration/handler/WebConfiguration.java
@@ -1,10 +1,9 @@
 package project.swithme.order.common.configuration.handler;
 
+import java.util.List;
 import org.springframework.stereotype.Component;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
 
 @Component
 public class WebConfiguration implements WebMvcConfigurer {
