+++ order/src/main/java/project/swithme/order/common/configuration/webclient/WebClientConfiguration.java
@@ -1,15 +1,13 @@
 package project.swithme.order.common.configuration.webclient;
 
+import static org.springframework.web.util.DefaultUriBuilderFactory.EncodingMode.VALUES_ONLY;
+import java.time.Duration;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.reactive.ReactorResourceFactory;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.util.DefaultUriBuilderFactory;
-
-import java.time.Duration;
-
-import static org.springframework.web.util.DefaultUriBuilderFactory.EncodingMode.VALUES_ONLY;
 
 @Configuration
 public class WebClientConfiguration {
@@ -30,8 +28,8 @@ public class WebClientConfiguration {
         DefaultUriBuilderFactory factory = new DefaultUriBuilderFactory(baseUrl);
         factory.setEncodingMode(VALUES_ONLY);
         return WebClient.builder()
-                .uriBuilderFactory(factory)
-                .baseUrl(baseUrl)
-                .build();
+            .uriBuilderFactory(factory)
+            .baseUrl(baseUrl)
+            .build();
     }
 }
