+++ order/src/main/java/project/swithme/order/core/domain/order/infrastructure/OrderJpaRepository.java
@@ -4,4 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import project.swithme.order.core.domain.order.entity.Order;
 
 public interface OrderJpaRepository extends JpaRepository<Order, Long> {
+
 }
