+++ order/src/main/java/project/swithme/order/core/domain/reservation/infrastructure/ReservationJpaRepository.java
@@ -4,4 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import project.swithme.order.core.domain.reservation.entity.Reservation;
 
 public interface ReservationJpaRepository extends JpaRepository<Reservation, Long> {
+
 }
