+++ order/src/main/java/project/swithme/order/core/web/order/application/command/OrderCreateCommand.java
@@ -1,22 +1,20 @@
 package project.swithme.order.core.web.order.application.command;
 
+import static project.swithme.order.core.domain.order.entity.OrderLine.createOrderLine;
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.List;
 import project.swithme.order.core.domain.order.entity.Order;
 import project.swithme.order.core.domain.order.entity.OrderLine;
 import project.swithme.order.core.domain.order.entity.PayType;
 
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.List;
-
-import static project.swithme.order.core.domain.order.entity.OrderLine.createOrderLine;
-
 public record OrderCreateCommand(
-        Long studyCafeId,
-        Long productId,
-        BigDecimal productPrice,
-        Long lockerId,
-        BigDecimal lockerPrice,
-        PayType payType
+    Long studyCafeId,
+    Long productId,
+    BigDecimal productPrice,
+    Long lockerId,
+    BigDecimal lockerPrice,
+    PayType payType
 ) {
 
     public OrderCreateCommand {
@@ -24,8 +22,8 @@ public record OrderCreateCommand(
     }
 
     private void validateLocker(
-            Long lockerId,
-            BigDecimal lockerPrice
+        Long lockerId,
+        BigDecimal lockerPrice
     ) {
         if (lockerId != null && lockerPrice == null) {
             throw new IllegalArgumentException("올바른 락커 정보를 입력해주세요.");
