+++ order/src/main/java/project/swithme/order/core/web/order/facade/validator/OrderValidator.java
@@ -13,11 +13,11 @@ public class OrderValidator {
 
     public void validate(OrderCreateCommand orderCreateCommand) {
         productClient.validateProductDetails(
-                orderCreateCommand.studyCafeId(),
-                orderCreateCommand.productId(),
-                orderCreateCommand.productPrice(),
-                orderCreateCommand.lockerId(),
-                orderCreateCommand.lockerPrice()
+            orderCreateCommand.studyCafeId(),
+            orderCreateCommand.productId(),
+            orderCreateCommand.productPrice(),
+            orderCreateCommand.lockerId(),
+            orderCreateCommand.lockerPrice()
         );
     }
 }
