+++ order/src/test/java/project/swithme/order/test/IntegrationTestBase.java
@@ -1,5 +1,8 @@
 package project.swithme.order.test;
 
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.modifyUris;
+import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+import static org.springframework.restdocs.restassured.RestAssuredRestDocumentation.documentationConfiguration;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -16,17 +19,10 @@ import org.springframework.restdocs.operation.preprocess.OperationPreprocessor;
 import org.springframework.restdocs.restassured.RestAssuredOperationPreprocessorsConfigurer;
 import project.swithme.order.common.annotation.IntegrationTest;
 
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.modifyUris;
-import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
-import static org.springframework.restdocs.restassured.RestAssuredRestDocumentation.documentationConfiguration;
-
 @IntegrationTest
 public abstract class IntegrationTestBase {
 
     private static final String SCHEMA = "www.study-with-me.com";
-
-    @Autowired
-    private RdbInitializationConfiguration databaseInitialization;
 
     @LocalServerPort
     protected int port;
@@ -36,6 +32,9 @@ public abstract class IntegrationTestBase {
 
     protected RequestSpecification specification;
 
+    @Autowired
+    private RdbInitializationConfiguration databaseInitialization;
+
     protected IntegrationTestBase() {
         initRestAssureConfiguration();
     }
@@ -43,30 +42,30 @@ public abstract class IntegrationTestBase {
     @BeforeEach
     void setUp(RestDocumentationContextProvider restDocumentation) {
         databaseInitialization.truncateAllEntity();
-        RestAssured.port = port;
 
         OperationPreprocessor operationPreprocessor = modifyUris()
-                .host(SCHEMA)
-                .removePort();
+            .host(SCHEMA)
+            .removePort();
 
-        RestAssuredOperationPreprocessorsConfigurer restDocumentationFilter = documentationConfiguration(restDocumentation)
-                .operationPreprocessors()
-                .withRequestDefaults(operationPreprocessor, prettyPrint())
-                .withResponseDefaults(prettyPrint());
+        RestAssuredOperationPreprocessorsConfigurer restDocumentationFilter = documentationConfiguration(
+            restDocumentation)
+            .operationPreprocessors()
+            .withRequestDefaults(operationPreprocessor, prettyPrint())
+            .withResponseDefaults(prettyPrint());
 
         this.specification = new RequestSpecBuilder()
-                .setPort(port)
-                .addFilter(restDocumentationFilter)
-                .build();
+            .setPort(port)
+            .addFilter(restDocumentationFilter)
+            .build();
     }
 
     private void initRestAssureConfiguration() {
         objectMapper = new ObjectMapper()
-                .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
+            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
         objectMapper.setSerializationInclusion(JsonInclude.Include.ALWAYS);
 
         RestAssured.config = new RestAssuredConfig().objectMapperConfig(
-                new ObjectMapperConfig().jackson2ObjectMapperFactory((clazz, charset) -> objectMapper)
+            new ObjectMapperConfig().jackson2ObjectMapperFactory((clazz, charset) -> objectMapper)
         );
     }
 }
