+++ order/src/test/java/project/swithme/order/test/RdbInitializationConfiguration.java
@@ -5,13 +5,12 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.metamodel.EntityType;
 import jakarta.persistence.metamodel.Metamodel;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 @Component
 public class RdbInitializationConfiguration {
@@ -23,29 +22,29 @@ public class RdbInitializationConfiguration {
     private Set<String> tableNames;
 
     public RdbInitializationConfiguration(
-            EntityManager entityManager,
-            Set<String> tableNames
+        EntityManager entityManager,
+        Set<String> tableNames
     ) {
         this.entityManager = entityManager;
         this.tableNames = tableNames;
-    }
-
-    @PostConstruct
-    public void afterPropertiesSet() {
-        Metamodel metamodel = entityManager.getMetamodel();
-        tableNames = metamodel.getEntities().stream()
-                .filter(isEntityTypeAndNotNull())
-                .map(toLowerCase())
-                .collect(Collectors.toUnmodifiableSet());
     }
 
     private static Function<EntityType<?>, String> toLowerCase() {
         return entityType -> entityType.getName().toLowerCase();
     }
 
+    @PostConstruct
+    public void afterPropertiesSet() {
+        Metamodel metamodel = entityManager.getMetamodel();
+        tableNames = metamodel.getEntities().stream()
+            .filter(isEntityTypeAndNotNull())
+            .map(toLowerCase())
+            .collect(Collectors.toUnmodifiableSet());
+    }
+
     private Predicate<EntityType<?>> isEntityTypeAndNotNull() {
         return entityType -> entityType.getJavaType()
-                .getAnnotation(Entity.class) != null;
+            .getAnnotation(Entity.class) != null;
     }
 
     @Transactional
