+++ order/src/test/java/project/swithme/order/test/order/domain/PayGroupUnitTest.java
@@ -1,12 +1,11 @@
 package project.swithme.order.test.order.domain;
 
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static project.swithme.order.core.domain.order.entity.PayGroup.findPayTypes;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import project.swithme.order.core.domain.order.entity.PayGroup;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static project.swithme.order.core.domain.order.entity.PayGroup.findPayTypes;
 
 @DisplayName("[UnitTest] 결제 그룹 단위 테스트")
 class PayGroupUnitTest {
