+++ order/src/test/java/project/swithme/order/test/order/integrationtest/OrderIntegrationTest.java
@@ -1,5 +1,7 @@
-package project.swithme.order.test.order.integration_test;
+package project.swithme.order.test.order.integrationtest;
 
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,10 +9,6 @@ import project.swithme.order.common.auth.StudyWithMeUser;
 import project.swithme.order.core.web.order.facade.OrderFacade;
 import project.swithme.order.core.web.order.presentation.request.OrderRequest;
 import project.swithme.order.test.IntegrationTestBase;
-
-import java.math.BigDecimal;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @DisplayName("[IntegrationTest] 주문 통합 테스트")
 class OrderIntegrationTest extends IntegrationTestBase {
@@ -23,12 +21,12 @@ class OrderIntegrationTest extends IntegrationTestBase {
     void order_create_test() {
         StudyWithMeUser studyWithMeUser = new StudyWithMeUser(1L);
         OrderRequest orderRequest = createOrderRequest(
-                1L,
-                1L,
-                BigDecimal.valueOf(75000),
-                1L,
-                BigDecimal.valueOf(10000),
-                "카카오 페이"
+            1L,
+            1L,
+            BigDecimal.valueOf(75000),
+            1L,
+            BigDecimal.valueOf(10000),
+            "카카오 페이"
         );
 
         Long newOrderId = orderFacade.createOrder(studyWithMeUser, orderRequest.toCommand());
@@ -36,20 +34,20 @@ class OrderIntegrationTest extends IntegrationTestBase {
     }
 
     private OrderRequest createOrderRequest(
-            Long studyCafeId,
-            Long productId,
-            BigDecimal productPrice,
-            Long lockerId,
-            BigDecimal lockerPrice,
-            String payType
+        Long studyCafeId,
+        Long productId,
+        BigDecimal productPrice,
+        Long lockerId,
+        BigDecimal lockerPrice,
+        String payType
     ) {
         return new OrderRequest(
-                studyCafeId,
-                productId,
-                productPrice,
-                lockerId,
-                lockerPrice,
-                payType
+            studyCafeId,
+            productId,
+            productPrice,
+            lockerId,
+            lockerPrice,
+            payType
         );
     }
 }
